Created the LSTMModel that treats all pipelines as if they're straight
